+++ components/src/Link/Link.js
@@ -24,7 +24,7 @@ const Link = styled.a(color, space, ({ underline, ...props }) => ({
   },
   "&:focus": {
     outline: "none",
-    boxShadow: `0 0 3px ${theme.colors.blue}`
+    boxShadow: theme.shadows.focus
   }
 }));