+++ docs/src/shared/menuData.js
@@ -13,6 +13,10 @@ export const menuData = [
       {
         name: "Spacing",
         href: "/style/spacing"
+      },
+      {
+        name: "Shadows",
+        href: "/style/shadows"
       }
     ]
   },
@@ -158,8 +162,8 @@ export const menuData = [
         href: "https://github.com/nulogy/design-system"
       },
       {
-        name: "Tokens",
-        href: "/tokens"
+        name: "Theme",
+        href: "/theme"
       },
       {
         name: "Reading list",